Rename file that manages database engines to avoid confusion
